requirements.txt
beautifulsoup4==4.4.1
chardet==2.3.0
decorator==4.0.6
gyp==0.1
html5lib==0.999
ipython==2.4.1
lxml==3.5.0
pbr==3.1.1
pexpect==4.0.1
ptyprocess==0.5
pycurl==7.43.0
Pygments==2.1
pyliblzma==0.5.3
pyserial==3.4
pysqlite==1.0.1
rpm-python==4.12.0.1
simplegeneric==0.8.1
six==1.11.0
stevedore==1.27.1
urlgrabber==3.9.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
virtualenvwrapper-win==1.2.2
yum-metadata-parser==1.1.4
